app/requirements_test.txt
# Testing dependencies for comprehensive test suite

# Core testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-timeout>=2.1.0
pytest-mock>=3.10.0
pytest-json-report>=1.5.0

# Coverage reporting
pytest-cov>=4.0.0

# Performance and load testing
pytest-benchmark>=4.0.0
pytest-xdist>=3.0.0  # For parallel test execution

# HTTP testing
aiohttp>=3.8.0
requests>=2.28.0
responses>=0.22.0

# Database testing
pytest-postgresql>=4.1.0  # For PostgreSQL testing
sqlalchemy>=1.4.0

# Mocking and fixtures
factory-boy>=3.2.0
freezegun>=1.2.0  # For time-based testing

# Utilities
psutil>=5.9.0  # For system resource monitoring
memory-profiler>=0.60.0  # For memory usage testing

# Existing dependencies (ensure compatibility)
psycopg2-binary
pydantic
python-dotenv
fastapi>=0.111.0
uvicorn[standard]>=0.30.1